Avoiding an unnecessary branch
Since if the percentiles aren't ready yet, we can simply prepare them and still run the tests afterwards. This avoids an unnecessary call to the prepare_inputs function (which could possibly take long, yet it's not that much of an optimization since the possibly many next calls will still be called.)

Note that the percentile are computed using the very first exec_times values and as such are not very accurate for the next tests.