bugfix: incorrect tau calculation
The raw t-statistic value is of little information if the number of samples
(measurements) is not provided. Hence, we try to provide also a normalized
version that can be used to compare t-values, even if they were taken with
different sample sizes.

The "normalized" t-statistic is called tau in the code. It is meant to be
independent of the sample size. Hence, a natural choice is to make
tau = t / sqrt(N).

The code contained a stupid bug that made tau depend on N (exactly the
opposite we wished for). This change fixes this.